+++ webAdopcion/src/main/java/fluffandpaws/webadopcion/Controllers/UserController.java
@@ -24,7 +24,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")//Esto nos retorna el usuario
-    public Usuarios getUsuarios(@PathVariable Integer id) {
+    public Usuarios getUsuarios(@PathVariable Long id) {
 
         return user.findById(id).orElseThrow();
     }
@@ -38,7 +38,7 @@ public class UserController {
 
     @PutMapping("/{id}")//para modificar usuario
 
-    public Usuarios replacePost(@PathVariable Integer id, @RequestBody Usuarios newUser) {
+    public Usuarios replacePost(@PathVariable Long id, @RequestBody Usuarios newUser) {
 
         newUser.setId(id);
         user.replace(newUser);
@@ -47,7 +47,7 @@ public class UserController {
     }
     @DeleteMapping("/{id})")//borramos el usuario
 
-    public Usuarios deleteUsuarios(@PathVariable Integer id){
+    public Usuarios deleteUsuarios(@PathVariable Long id){
 
         Usuarios u = user.findById(id).orElseThrow();
         user.deleteById(id);
