+++ webAdopcion/src/main/java/fluffandpaws/webadopcion/Repositories/MessagesRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MessagesRepository  extends CrudRepository<Messages, Integer>
+public interface MessagesRepository  extends CrudRepository<Messages, Long>
 {
