+++ webAdopcion/src/main/java/fluffandpaws/webadopcion/Repositories/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 //Spring lo implementa automáticamente
 @Repository
-public interface UserRepository extends CrudRepository<Usuarios, Integer> {
+public interface UserRepository extends CrudRepository<Usuarios, Long> {
     //CRUD = Create, Read, Update, Delete
 }
